app/api-rest/requirements.txt
boto3==1.9.91
Flask==1.0.2
flask-restplus==0.12.1
Flask-Script==2.0.6
python-etcd==0.4.5
gunicorn==19.9.0


